[US_ND] Remove folks on solitary confinement from both facilities opps (Recidiviz/recidiviz-data#35922)
## Description of the change

[US_ND] Remove folks on solitary confinement (BIU) from both facilities
opps.

I uploaded it to a sandbox under (_hsalas) and the changes are small and
in the right direction.

![Screenshot 2024-12-09 at 4 26
47 PM](https://github.com/user-attachments/assets/736b0002-438b-4a53-afba-b782c834da27)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c798d791b0eb349156e8be8659483565381064a4