[ID perf] query supervisors and officers; generate spreadsheets based on them (Recidiviz/recidiviz-data#35981)
## Description of the change

A few changes to support querying supervisors + officers to generate
sheets per officer:
* Added a file with some helper classes and functionality to read
supervisor+officer data from BigQuery
* Moved everything into a new folder so we can more cleanly have the
functionality spread between a few files
* Actually add the logic that creates multiple spreadsheets and names
the sheet titles appropriately

Ran it locally and it generated 31 spreadsheets, which matches the
number of supervisors shown in ID Insights. I spot checked a few of them
and the officers matched what we show in insights. Example screenshot:
https://drive.google.com/file/d/1toTHDGE7S6jU3cWx0hyViZk_PZ_cQYnu/view?usp=sharing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35928

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2b2702f1b10a829da347e2d730234d15eb118f54