[ID perf] add average daily population to ID perf spreadsheets (Recidiviz/recidiviz-data#35987)
## Description of the change

Adds a new `OfficerAggregatedMetrics` class with some relevant helpers
for reading aggregated metrics. Adds logic to the generate script to
call that class and write the data into the sheet.

Tested by running the script and spot checking a few of the yearly
values against what it says in Insights. Screenshot:
https://drive.google.com/file/d/13DU76vlT64n-De_X0r_e12lw2iX4QejR/view?usp=sharing.
Note that the Dec 2024 column will be empty until we have December data,
which won't be until Jan 1. For the same reason, I'm pretending the year
is Dec '23 - Dec '24, and will update the actual date in January.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b60bc19e27a89e3556f3a1e394fe74e0aace7e5b