[Justice Counts] Remove excessive logging from Bulk Upload (Recidiviz/recidiviz-data#35977)
## Description of the change

This PR removes an outdated logging message related to agency uploads.
The current logging message is incorrect, as it suggests there are
errors and data was only partially saved, which is misleading.

The logging was added during a previous refactor to log every time an
agency uploads a sheet that isn't perfect, which provides limited value
and potentially clutters the logs.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ac32061671dbfdf5a9e2bce0ceca29d2807caa2a