[Insights][Action Strategies] Add timestamp into unique constraint on action_strategy_surfaced_events (Recidiviz/recidiviz-data#36015)
## Description of the change

Adds `timestamp` into the unique constraint on
`action_strategy_surfaced_events`. One of the action strategies can be
surfaced more than once, so we will need to allow the same
strategy/supervisor/officer event to be inserted multiple times. We saw
an error pop up in production today because this unique constraint was
tripped for a legitimate event.

I had considered just dropping the constraint altogether, but we DONT
want multiple events per day so I figured I'd leave the constraint on
there so that we are alerted if there is more than one of the same type
per day.

I ran this migration in my local database(s) and loaded the updated data
into them.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#36007

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c6369005dff26f83b63b35daa545df1277666050