[Insights][Action Strategies] Respond with 200 for leadership users on get_action_strategies (Recidiviz/recidiviz-data#36043)
## Description of the change

Respond with 200 and "empty" response body for leadership users who are
not supervisors.

We have a bug where leadership users are receiving 401 on the
`get_action_strategies` endpoint which is crashing the page. We don't
have any specific requirements around how to handle leadership users, so
I propose that we send an "empty" object instead of a 401. There may be
a case where some leadership users are also supervisors, so not
initiating the request in the FE for leadership users seems like the
wrong move.

The response in this PR is not truly empty, it is: `{
supervisor_pseudo_id: None }`. I had considered just sending an empty
object, but this seems more explicit. I'm happy to change this if folks
have opinions.

I have a FE PR coming soon that adds tests to ensure that this is
compatible in the FE.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#36040

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 511b56ad6c160b01296fd4c9fb593e460478b0fb