Fix approach to formatting in `line_labels()` (Recidiviz/recidiviz-data#36072)
## Description of the change

Updated the `line_labels()` function so that labels are not reformatted
by default. (This is the `line_labels()` counterpart to [this previous
PR](https://github.com/Recidiviz/recidiviz-data/pull/35797) that updated
the formatting in `add_legend()`.)

Now, it is necessary to specify `format_labels=True` in order to have
any reformatting. The default function for reformatting remains the
same, and users should still be able to provide an alternative function
for reformatting if desired.

Existing code for plots should still run without errors, but line labels
will look different. DAs (and anyone else who uses the plotting
functions) would need to update their code to now explicitly opt in to
reformatting if desired, so folks may want to double-check their
notebooks and make sure their line labels will still be set up in the
way they'd like.

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c730d18b4ec20cca00a630da3f3b8000b369a9c3