[Workflows] Fix typing bug preventing Workflows config promotion (Recidiviz/recidiviz-data#36062)
Fix a typing issue that was causing errors in the process of promoting certain Workflows opportunity configs. Add real data to tests to make this sort of bug more difficult to introduce in the future. Two fields had included values annotated as `list[str]` instead of `str`, which caused the `OpportunityConfig` object's `to_dict` method to silently convert these fields to lists.

---------

Co-authored-by: Daniel McLaughlin <danielm@recidiviz.org>
GitOrigin-RevId: d10fc7e02dbbe159b6f94a6c86d2d498cd19a9c8