[Vitals] add supervision_*_30_day_aggregated_metrics (Recidiviz/recidiviz-data#36020)
## Description of the change

Adds aggregated metrics tables that calculate a 30 day rate for all the
same metrics / units of analysis as the "day" tables. This will be used
in the legacy vitals queries.

Since it's only calculating 30 days worth of rates, it doesn't take
super long to load- about 17.78 seconds locally, which is comparable to
the single day version, so I feel comfortable checking this in now
instead of waiting for the perf improvements.

I also took this opportunity to limit which states we include in all the
vitals agg metrics tables, since there's not much reason to include
states that we aren't doing vitals work in.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34507

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: dbe8bdd6a812b030dd6e2b05899e2401322d81c5